Most Used Tags, TodoItemDto Colouring
Most Used Tags are statically generated and located on To-do List page.
TodoItemDto has a color property now and to-do list item is coloured by this value.
TodoItemDto has a tags array now and it is generated during initialization of the object.
TodoItemDto editor has a color defining option and "adding or removing tags" functionality.